requirements.txt
streamlit
torch
transformers
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
pandas
requests
pytz
psycopg2-binary
groq
python-dateutil
sendgrid
